==> avbernat/Dispersal/Autumn_2019/windaq_processing/0.demographics.py <==
import os
import csv

#***************************************************************************************
# Updating the all_dispersal_data sheet with sex, site, and host plant.
#***************************************************************************************

git_path = r"/Users/anastasiabernat/Desktop/git_repositories/SBB-dispersal/"
data_path = "avbernat_working_on/Dispersal/Autumn_2019/windaq_processing/data/"

demographics_data = git_path + data_path + "1.bug_demographics_data_coor.csv"
all_dispersal_data = git_path + data_path +"1.dispersal_data.csv"

sex_dict = {} 
site_dict = {}
host_dict = {}
lat_dict = {}
long_dict = {}

with open(demographics_data, "r") as demo_data:
    reader = csv.DictReader(demo_data)
    for row in reader:
        ID = row["ID"]
        sex = row["sex"]
        pop = row["population"]
        host = row["host"]
        site = row["site"]
        lat = row["latitude"]
        long = row["longitude"]

        if ID not in sex_dict:
            sex_dict[ID] = sex
        if (ID, pop) not in site_dict:
            site_dict[(ID, pop)] = site
        if (ID, site) not in host_dict:
            host_dict[(ID, site)] = host
        if (ID, site) not in lat_dict:
            lat_dict[(ID, site)] = lat
        if lat not in long_dict:
            long_dict[lat] = long

full_data = [] 
with open(all_dispersal_data, "r") as all_data:
    reader = csv.DictReader(all_data)
    for r in reader:
        row_data = {}
        ID_num = r["ID"]
        population = r["population"]
        sex = sex_dict[ID_num]
        site = site_dict[(ID_num, population)]
        host_plant = host_dict[(ID_num, site)]
        lat = lat_dict[(ID_num, site)]
        long = long_dict[lat]
        
        
        row_data["ID"] = ID_num
        row_data["box"] = r["box"]
        row_data["test_date"] = r["test_date"]
        row_data["time_start"] = r["time_start"]
        row_data["set_number"] = r["set_number"].lstrip('0')
        row_data["chamber"] = r["chamber"]
        row_data["sex"] = sex
        row_data["population"] = population
        row_data["site"] = site
        row_data["host_plant"] = host_plant.rstrip()
        row_data["flew"] = r["flew"]
        row_data["died?"] = r["died?"]
        row_data["flight_type"] = r["flight_type"]
        row_data["flight_details"] = r["flight_details"]
        row_data["NOTES"] = r["NOTES"]
        row_data["filename"] = r["filename"]
        row_data["mass"] = r["mass"]
        row_data["short-wing?"] = r["short-wing?"]
        row_data["eggs"] = r["eggs"]
        row_data["time_end"] = r["time_end"]
        row_data["latitude"] = lat
        row_data["longitutde"] = long


        full_data.append(row_data)

#print(full_data[0:5])

outpath = git_path + data_path + "1.all_dispersal_data.csv"

with open(outpath, "w") as output_file:
    writer = csv.DictWriter(output_file, fieldnames = row_data.keys())
    writer.writeheader()
    for r in full_data:
        writer.writerow(r)
